backend/requirements.txt
Django==5.2.6
psycopg2-binary
djangorestframework
drf-spectacular
django-cors-headers